Moved `NSPersistentContainer` extension to separate file
