+++ FeedStoreChallenge/CoreData/CoreDataFeedStore.swift
@@ -86,38 +86,6 @@ public final class CoreDataFeedStore: FeedStore {
 	}
 }
 
-private extension NSPersistentContainer {
-	
-	enum LoadingError: Error {
-		case modelNotFound
-		case cannotLoadPersistentStores(Error)
-	}
-	
-	static func load(dataModelName: String, storeURL: URL, in bundle: Bundle) throws -> NSPersistentContainer {
-		guard let model = NSManagedObjectModel(name: dataModelName, in: bundle) else {
-			throw LoadingError.modelNotFound
-		}
-		
-		let container = NSPersistentContainer(name: dataModelName, managedObjectModel: model)
-		let description = NSPersistentStoreDescription(url: storeURL)
-		container.persistentStoreDescriptions = [description]
-		var loadError: Error?
-		container.loadPersistentStores { loadError = $1 }
-		try loadError.map { throw LoadingError.cannotLoadPersistentStores($0) }
-		
-		return container
-	}
-}
-
-private extension NSManagedObjectModel {
-	convenience init?(name: String, in bundle: Bundle) {
-		guard let modelURL = bundle.url(forResource: name, withExtension: "momd") else {
-			return nil
-		}
-		self.init(contentsOf: modelURL)
-	}
-}
-
 private extension CDFeed {
 	var fetchRequest: NSFetchRequest<CDFeed> {
 		NSFetchRequest<CDFeed>(entityName: CDFeed.entity().name!)
